AllResultController:코드 분리 후 주석 처리 / ResultController / ResultService / ResultServiceImpl
